refactor: extract BaseCompiler and consolidate prompts (#674, #678)
Extract shared compiler infrastructure (constructor, _call_llm,
_generate_unique_slug, _index_article, _create_synthesizes_links) into
BaseCompiler base class, eliminating ~345 lines of duplication across
compiler.py, concept_compiler.py, and synthesis_compiler.py.

Consolidate prompt templates into engine/prompts.py for single-source
management. Update test mock patch paths to target base_compiler module
where get_llm_router and get_settings are now imported.
