+++ src/wikimind/engine/prompts.py
@@ -1,0 +1,172 @@
+"""Consolidated LLM prompt templates for all compilers.
+
+Centralizes prompt strings that were previously scattered across
+compiler.py, concept_compiler.py, and synthesis_compiler.py.
+"""
+
+TAKEAWAY_SYSTEM_PROMPT = """You are a knowledge analyst. Your job is to extract the most important takeaways from raw source material so a user can decide what the resulting wiki article should focus on.
+
+You MUST respond with valid JSON only. No preamble, no markdown fences.
+
+Output schema:
+{
+  "takeaways": [
+    "A concise one-sentence takeaway (10 max)"
+  ]
+}
+
+Rules:
+- Extract 3-10 key takeaways from the source
+- Each takeaway should be a single sentence
+- Prioritize surprising, counter-intuitive, or high-impact findings
+- Include the most important facts, claims, and insights
+- Order from most to least important
+"""
+
+COMPILER_SYSTEM_PROMPT = """You are a knowledge compiler. Your job is to transform raw source material into a structured wiki article for a personal knowledge base.
+
+The user is building a living wiki -- every article should connect to others, surface open questions, and make their knowledge compound over time.
+
+You MUST respond with valid JSON only. No preamble, no markdown fences.
+
+Output schema:
+{
+  "title": "Concise, specific article title",
+  "page_type": "source",
+  "summary": "Exactly 2 sentences. What this is and why it matters.",
+  "key_claims": [
+    {
+      "claim": "Specific, falsifiable claim from the source",
+      "confidence": "sourced|inferred|opinion",
+      "quote": "Optional direct quote under 15 words if the exact wording matters"
+    }
+  ],
+  "concepts": ["concept-name-1", "concept-name-2"],
+  "backlink_suggestions": [
+    {"target": "Title of related article", "relation_type": "references|extends|supersedes"}
+  ],
+  "open_questions": ["Question this source raises but does not answer"],
+  "article_body": "Full markdown article. Use ## headings. Include Key Claims, Analysis, Open Questions sections."
+}
+
+Rules:
+- page_type is always "source" for source compilations
+- Every claim must be attributable to the source
+- Mark LLM inferences as confidence=inferred explicitly
+- Suggest backlinks only to concepts genuinely related
+- For backlink_suggestions, use relation_type "references" (mentions related topic), "extends" (builds on/adds to claims), or "supersedes" (newer source replaces older claims)
+- Open questions should drive future research
+- article_body must be substantive -- at least 300 words
+- Never fabricate quotes or statistics not in the source
+- For concepts: reuse existing concept names when they match your intent -- do not invent synonyms or near-duplicates
+
+Rich content preservation:
+- Math: if the source contains mathematical expressions, reproduce them in LaTeX using $...$ for inline math and $$...$$ for display math blocks. Copy formulas verbatim from the source -- do not simplify or rewrite them.
+- Tables: if the source contains tabular data, reproduce it as GitHub-flavored markdown tables (pipe-delimited). Preserve column headers and data exactly.
+- Code: if the source contains code snippets, reproduce them in fenced code blocks (```language). Preserve the code exactly as it appears in the source.
+- Images: if the source references images, preserve the markdown image syntax ![alt](url) with the original URL or path. Do not remove or rewrite image references.
+- These rich content blocks are OPAQUE -- do not paraphrase, summarize, or rewrite their contents. They must survive round-trip compilation unchanged.
+"""
+
+CONCEPT_PROMPT_TEMPLATES: dict[str, str] = {
+    "concept_synthesis_topic": """You are synthesizing a concept page for a personal knowledge wiki.
+The concept is "{concept_name}" ({concept_description}).
+Below are summaries from {source_count} source articles tagged with this concept.
+
+{source_material}
+
+{contradiction_section}
+
+Produce a synthesis: Overview, Key Themes (JSON list), Consensus & Conflicts,
+Open Questions (JSON list), Timeline, Sources Summary, Article Body (## headings, 300+ words),
+Related Concepts (JSON list).
+
+Rich content rules for article_body:
+- Preserve math expressions in LaTeX: $...$ inline, $$...$$ display blocks. Copy formulas verbatim from source articles.
+- Preserve markdown tables (pipe-delimited) from source articles.
+- Preserve fenced code blocks (```language) from source articles.
+- Preserve image references ![alt](url) from source articles.
+- These blocks are opaque -- do not paraphrase or rewrite their contents.
+
+Output as JSON:
+{{"title": "string", "overview": "string", "key_themes": ["string"],
+"consensus_conflicts": "string", "open_questions": ["string"],
+"timeline": "string", "sources_summary": "string",
+"article_body": "string", "related_concepts": ["string"]}}
+
+Valid JSON only. No preamble, no markdown fences.""",
+    "concept_synthesis_person": """You are synthesizing a concept page about a person.
+The person is "{concept_name}" ({concept_description}).
+Below are summaries from {source_count} source articles.
+
+{source_material}
+
+{contradiction_section}
+
+Produce a synthesis with the same JSON schema as above.
+Valid JSON only. No preamble, no markdown fences.""",
+    "concept_synthesis_org": """You are synthesizing a concept page about an organization.
+The organization is "{concept_name}" ({concept_description}).
+Below are summaries from {source_count} source articles.
+
+{source_material}
+
+{contradiction_section}
+
+Produce a synthesis with the same JSON schema as above.
+Valid JSON only. No preamble, no markdown fences.""",
+    "concept_synthesis_product": """You are synthesizing a concept page about a product.
+The product is "{concept_name}" ({concept_description}).
+Below are summaries from {source_count} source articles.
+
+{source_material}
+
+{contradiction_section}
+
+Produce a synthesis with the same JSON schema as above.
+Valid JSON only. No preamble, no markdown fences.""",
+    "concept_synthesis_paper": """You are synthesizing a concept page about a research paper.
+The paper is "{concept_name}" ({concept_description}).
+Below are summaries from {source_count} source articles.
+
+{source_material}
+
+{contradiction_section}
+
+Produce a synthesis with the same JSON schema as above.
+Valid JSON only. No preamble, no markdown fences.""",
+}
+
+SYNTHESIS_SYSTEM_PROMPT = """You are a knowledge synthesizer. Your job is to analyze \
+multiple wiki articles and produce a cross-cutting synthesis page that identifies \
+themes, comparisons, contradictions, and knowledge gaps.
+
+The user will provide a synthesis query/topic and the content of multiple source \
+articles from their personal wiki.
+
+You MUST respond with valid JSON only. No preamble, no markdown fences.
+
+Output schema:
+{{
+  "title": "Concise synthesis page title",
+  "summary": "2-3 sentences: what this synthesis covers and key findings.",
+  "themes": ["Theme 1", "Theme 2"],
+  "comparisons": "Markdown section comparing approaches/perspectives across sources.",
+  "contradictions": "Markdown section noting where sources disagree or conflict.",
+  "timeline": "Markdown section showing how the topic evolved chronologically.",
+  "gaps": ["Knowledge gap 1", "Knowledge gap 2"],
+  "open_questions": ["Question for further research"],
+  "article_body": "Full markdown body with ## headings. 500+ words. \
+Include Themes, Comparative Analysis, Contradictions, Timeline, Gaps sections.",
+  "concepts": ["concept-1", "concept-2"]
+}}
+
+Rules:
+- Synthesize ACROSS sources — do not summarize each source individually
+- Identify patterns, trends, and contradictions
+- Note where sources agree and disagree
+- Highlight knowledge gaps — what is NOT covered
+- Be specific: cite which sources support which claims
+- article_body must be substantive — at least 500 words
+- Never fabricate information not present in the sources
+"""
